core/requirements.txt
django
djangorestframework
psycopg2-binary
stripe
python-dotenv
django-environ
drf-spectacular
Pillow
djangorestframework-simplejwt
django-debug-toolbar
google-generativeai